+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/PainterContentScaleTestScreen.android.kt
@@ -1,6 +1,7 @@
 package com.github.panpf.sketch.sample.ui.test
 
 import android.os.Build
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
@@ -20,7 +21,8 @@ import com.github.panpf.sketch.util.toScale
 
 actual suspend fun buildPainterContentScaleTestPainters(
     context: PlatformContext,
-    contentScale: ContentScale
+    contentScale: ContentScale,
+    alignment: Alignment,
 ): List<Pair<String, Painter>> {
     val list = mutableListOf<Pair<String, Painter>>()
 
@@ -40,7 +42,8 @@ actual suspend fun buildPainterContentScaleTestPainters(
     ImageRequest(context, ResourceImages.clockHor.uri).execute()
         .let { it as ImageResult.Success }
         .image.asPainter().apply {
-            val crossfadePainter = CrossfadePainter(null, this, contentScale = contentScale)
+            val crossfadePainter =
+                CrossfadePainter(null, this, contentScale = contentScale, alignment = alignment)
             list.add("CrossfadePainter\nImageBitmapPainter\nImageBitmap" to crossfadePainter)
         }