+++ sketch-compose-core/src/commonTest/kotlin/com/github/panpf/sketch/compose/core/common/test/painter/CrossfadePainterTest.kt
@@ -60,6 +60,8 @@ class CrossfadePainterTest {
             assertFalse(fadeStart)
             assertTrue(preferExactIntrinsicSize)
         }
+
+        // TODO teest alignment
     }
 
     @Test