+++ sketch-compose-core/src/commonTest/kotlin/com/github/panpf/sketch/compose/core/common/test/transition/ComposeCrossfadeTransitionTest.kt
@@ -92,6 +92,8 @@ class ComposeCrossfadeTransitionTest {
         assertFailsWith(IllegalArgumentException::class) {
             ComposeCrossfadeTransition(sketch, request, imageViewTarget, result, durationMillis = 0)
         }
+
+        // TODO teest alignment
     }
 
     @Test